RavenDB-25281 Split QueryPlanBuilder into three partial files
Structural (parsing, plan emission), per-execution (resolution,
sorting, spatial/vector), and diagnostic (inspection graph).
